requirements.txt
websocket-client==1.5.3
